src/requirements.txt
asgiref==3.5.0; python_version >= "3.8"
django-cors-headers==3.11.0; python_version >= "3.7"
django-timezone-field==5.0; python_version >= "3.7" and python_version < "4.0"
django==4.0.2; python_version >= "3.8"
djangorestframework==3.13.1; python_version >= "3.6"
pytz==2021.3; python_version >= "3.7" and python_version < "4.0"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
